requirements.txt
appnope==0.1.3
asttokens==2.4.0
attrs @ file:///private/tmp/python-attrs-20231231-5195-5kv3i2/attrs-23.2.0
backcall==0.2.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
comm==0.1.4
debugpy==1.8.0
decorator==5.1.1
executing==2.0.0
google-api-core==2.12.0
google-api-python-client==2.104.0
google-auth==2.23.3
google-auth-httplib2==0.1.1
googleapis-common-protos==1.61.0
httplib2==0.22.0
idna==3.4
ipykernel==6.25.2
ipython==8.16.1
jedi==0.19.1
Jinja2 @ file:///private/tmp/python-jinja-20240110-4783-r651qg/Jinja2-3.1.3
jupyter_client==8.4.0
jupyter_core==5.4.0
MarkupSafe @ file:///private/tmp/python-markupsafe-20240119-4778-tstdmr/MarkupSafe-2.1.4
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
packaging @ file:///private/tmp/python-packaging-20231018-5368-zv724j/packaging-23.2
parso==0.8.3
pathspec @ file:///private/tmp/python-pathspec-20231212-4946-k7fih8/pathspec-0.12.1
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.11.0
prompt-toolkit==3.0.39
protobuf==4.24.4
psutil @ file:///private/tmp/python-psutil-20240119-4673-f71z61/psutil-5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser @ file:///private/tmp/pycparser-20231013-5787-1anpge2/pycparser-2.21
Pygments @ file:///private/tmp/pygments-20231121-5274-1rlyumb/pygments-2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pyzmq==25.1.1
requests==2.31.0
rsa==4.9
six==1.16.0
stack-data==0.6.3
tornado==6.3.3
traitlets==5.11.2
typing_extensions @ file:///private/tmp/python-typing-extensions-20231212-4978-cvkltg/typing_extensions-4.9.0
uritemplate==4.1.1
urllib3==2.0.7
wcwidth==0.2.8
